pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pydf"
version = "1.0.0"
description = "Biblioteca Python leve para extrair metadados de faturas em PDF, exportar para Excel e opcionalmente persistir em MySQL."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Daniel Arndt"}]
dependencies = [
  "openpyxl>=3.1.0",
  "pdfplumber>=0.11.0",
  "mysql-connector-python>=9.0.0",
]
keywords = ["pdf", "invoice", "excel", "mysql", "automation"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

[project.urls]
Documentation = "https://github.com/DanielArndt0/pydf/README.md"
Homepage = "https://github.com/DanielArndt0/pydf"
Repository = "https://github.com/DanielArndt0/pydf"
Issues = "https://github.com/DanielArndt0/pydf/issues"

[project.scripts]
pydf = "pydf.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "build>=1.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
